Refactor -  create 'pairing.ts' module for invite/accept functions
